+++ tests/test_healpix.py
@@ -173,3 +173,63 @@ def test_latlon_cuda_set_device_regression():
         grid.lat
     finally:
         torch.set_default_device(default)
+
+
+def test_zonal_average():
+    # hpx 2 in ring order
+    x = np.array(
+        [
+            0,
+            0,
+            0,
+            0,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            2,
+            2,
+            2,
+            2,
+            2,
+            2,
+            2,
+            2,
+            3,
+            3,
+            3,
+            3,
+            3,
+            3,
+            3,
+            3,
+            4,
+            4,
+            4,
+            4,
+            4,
+            4,
+            4,
+            4,
+            5,
+            5,
+            5,
+            5,
+            5,
+            5,
+            5,
+            5,
+            6,
+            6,
+            6,
+            6,
+        ]
+    )
+    x = x[None]
+    zonal = healpix.zonal_average(x)
+    assert zonal.shape == (1, 7)
+    assert np.all(zonal == np.arange(7))